Split agreement viewsets into domain-specific modules
