+++ agreements/api/viewsets/asset_agreement.py
@@ -5,14 +5,13 @@ from django.utils.timezone import now
 from rest_framework import viewsets
 from rest_framework.decorators import action
 from rest_framework.response import Response
-from agreements.api.serialziers.agreement_coverage import AgreementCoverageSerializer, AgreementCoverageWriteSerializer
-from agreements.api.serialziers.agreement_history import AgreementHistorySerializer, AgreementItemHistorySerializer
+from agreements.api.serialziers.agreement_coverage import AgreementCoverageSerializer
+from agreements.api.serialziers.agreement_history import AgreementHistorySerializer
 from agreements.api.serialziers.asset_agreement import AssetAgreementSerializer, AssetAgreementWriteSerializer
-from agreements.models.agreements import AgreementCoverage, AgreementHistory, AgreementItemHistory, AssetAgreement, AssetAgreementItem, CoverageScopeType
-from core.mixins import AuditMixin, ScopeFilterMixin
+from agreements.models.agreements import  AgreementHistory, AssetAgreement,  CoverageScopeType
+from core.mixins import ScopeFilterMixin
 from core.pagination import FlexiblePagination
-from agreements.api.serialziers.agreement_item import AssetAgreementItemSerializer, AssetAgreementItemWriteSerializer, resolve_asset_by_public_id
-from agreements.services.coverage import can_attach_asset_to_agreement
+from agreements.api.serialziers.agreement_item import AssetAgreementItemSerializer, resolve_asset_by_public_id
 from assets.models.assets import Accessory, Consumable, Equipment
 
 
@@ -445,454 +444,3 @@ class AssetAgreementViewSet( ScopeFilterMixin, viewsets.ModelViewSet, ):
         return Response(
             serializer.data
         )    
-
-class AgreementCoverageViewSet( ScopeFilterMixin, viewsets.ModelViewSet, ):
-
-    queryset = (
-        AgreementCoverage.objects
-        .select_related(
-            "agreement",
-            "department",
-            "location",
-            "room",
-        )
-        .order_by("id")
-    )
-
-    # permission_classes = [AssetAgreementPermission]
-
-    pagination_class = FlexiblePagination
-
-    lookup_field = "public_id"
-
-    def get_serializer_class(self):
-        if self.action in [
-            "create",
-            "update",
-            "partial_update",
-        ]:
-            return AgreementCoverageWriteSerializer
-        return AgreementCoverageSerializer
-    
-       # -------------------------
-    # Create
-    # -------------------------
-
-    def create(self, request, *args, **kwargs):
-
-        serializer = self.get_serializer(
-            data=request.data,
-            context={
-                "request": request
-            },
-        )
-
-        serializer.is_valid(
-            raise_exception=True
-        )
-
-        coverage = serializer.save()
-
-        # --------------------------------
-        # Resolve Scope Label
-        # --------------------------------
-
-        scope_target = (
-            coverage.department
-            or coverage.location
-            or coverage.room
-            or "GLOBAL"
-        )
-
-        # --------------------------------
-        # Audit Log
-        # --------------------------------
-
-        self.audit(
-            AuditLog.Events.AGREEMENT_COVERAGE_CREATED,
-            target=coverage,
-            description=(
-                f"{request.user.email} added "
-                f"{coverage.scope_type} coverage "
-                f"to agreement "
-                f"{coverage.agreement.public_id}"
-            ),
-            metadata={
-                "agreement_public_id":
-                    coverage.agreement.public_id,
-
-                "agreement_name":
-                    coverage.agreement.name,
-
-                "coverage_public_id":
-                    coverage.public_id,
-
-                "scope_type":
-                    coverage.scope_type,
-
-                "scope_target":
-                    str(scope_target),
-
-                "performed_by":
-                    request.user.email,
-            },
-        )
-
-        response_serializer = (
-            AgreementCoverageSerializer(
-                coverage,
-                context={
-                    "request": request
-                },
-            )
-        )
-
-        return Response(
-            response_serializer.data,
-            status=status.HTTP_201_CREATED,
-        )
-
-    # -------------------------
-    # Destroy
-    # -------------------------
-
-    def destroy( self, request, *args, **kwargs, ):
-
-        coverage = self.get_object()
-
-        scope_target = (
-            coverage.department
-            or coverage.location
-            or coverage.room
-            or "GLOBAL"
-        )
-
-        # --------------------------------
-        # Audit Log
-        # --------------------------------
-
-        self.audit(
-            AuditLog.Events.AGREEMENT_COVERAGE_REMOVED,
-            target=coverage,
-            description=(
-                f"{request.user.email} removed "
-                f"{coverage.scope_type} coverage "
-                f"from agreement "
-                f"{coverage.agreement.public_id}"
-            ),
-            metadata={
-                "agreement_public_id":
-                    coverage.agreement.public_id,
-                "agreement_name":
-                    coverage.agreement.name,
-                "coverage_public_id":
-                    coverage.public_id,
-                "scope_type":
-                    coverage.scope_type,
-                "scope_target":
-                    str(scope_target),
-                "performed_by":
-                    request.user.email,
-            },
-        )
-
-        coverage.delete()
-
-        return Response(
-            status=status.HTTP_204_NO_CONTENT,
-        )
-
-
-class AssetAgreementItemViewSet(AuditMixin, ScopeFilterMixin, viewsets.GenericViewSet, ):
-
-    queryset = (
-        AssetAgreementItem.objects
-        .select_related(
-            "agreement",
-            "equipment",
-            "consumable",
-            "accessory",
-        )
-        .order_by("id")
-    )
-
-    pagination_class = FlexiblePagination
-
-    lookup_field = "public_id"
-
-    # -------------------------
-    # Serializer Selection
-    # -------------------------
-
-    def get_serializer_class(self):
-        if self.action == "attach":
-            return AssetAgreementItemWriteSerializer
-        return AssetAgreementItemSerializer
-
-    # -------------------------
-    # List
-    # -------------------------
-
-    def list(self, request):
-
-        queryset = self.filter_queryset( self.get_queryset() )
-
-        page = self.paginate_queryset( queryset )
-
-        serializer = self.get_serializer(
-            page if page is not None else queryset,
-            many=True,
-            context={
-                "request": request
-            },
-        )
-
-        if page is not None:
-            return self.get_paginated_response(
-                serializer.data )
-
-        return Response(
-            serializer.data
-        )
-
-    # -------------------------
-    # Retrieve
-    # -------------------------
-
-    def retrieve( self, request, public_id=None, ):
-
-        item = self.get_object()
-        serializer = self.get_serializer(
-            item,
-            context={
-                "request": request
-            },
-        )
-
-        return Response( serializer.data )
-
-    # -------------------------
-    # Attach Asset
-    # -------------------------
-
-    @action(
-        detail=False,
-        methods=["post"],
-    )
-    def attach(self, request):
-
-        serializer = self.get_serializer(
-            data=request.data,
-            context={
-                "request": request
-            },
-        )
-
-        serializer.is_valid(
-            raise_exception=True
-        )
-
-        validated_data = (
-            serializer.validated_data
-        )
-
-        agreement = validated_data[
-            "agreement"
-        ]
-
-        # --------------------------------
-        # Resolve Asset
-        # --------------------------------
-
-        asset = (
-            validated_data.get(
-                "equipment"
-            )
-            or
-            validated_data.get(
-                "accessory"
-            )
-            or
-            validated_data.get(
-                "consumable"
-            )
-        )
-
-        # --------------------------------
-        # Eligibility Check
-        # --------------------------------
-
-        if not can_attach_asset_to_agreement(
-            agreement=agreement,
-            asset=asset,
-        ):
-
-            raise ValidationError(
-                {
-                    "non_field_errors": [
-                        (
-                            "This asset does not "
-                            "fall within the "
-                            "agreement coverage "
-                            "scope."
-                        )
-                    ]
-                }
-            )
-
-        # --------------------------------
-        # Save
-        # --------------------------------
-
-        item = serializer.save()
-
-        response_serializer = (
-            AssetAgreementItemSerializer(
-                item,
-                context={
-                    "request": request
-                },
-            )
-        )
-        self.audit(
-            AuditLog.Events.AGREEMENT_ITEM_ATTACHED,
-            target=item,
-            description=(
-                f"Attached asset "
-                f"{asset.public_id} "
-                f"to agreement "
-                f"{agreement.public_id}"
-            ),
-            metadata={
-                "agreement_public_id":
-                    agreement.public_id,
-                "agreement_name":
-                    agreement.name,
-                "asset_public_id":
-                    asset.public_id,
-                "asset_name":
-                    getattr(asset, "name", ""),
-                "asset_type":
-                    item.asset_type,
-                "agreement_item_public_id":
-                    item.public_id,
-                "performed_by":
-                    request.user.email,
-            },
-        )
-        
-
-        return Response(
-            response_serializer.data,
-            status=status.HTTP_201_CREATED,
-        )
-
-    # -------------------------
-    # Detach Asset
-    # -------------------------
-
-    @action(
-        detail=True,
-        methods=["post"],
-    )
-    def detach(
-        self,
-        request,
-        public_id=None,
-    ):
-
-        item = self.get_object()
-
-        asset = item.asset
-
-        agreement = item.agreement
-
-        # --------------------------------
-        # Audit Log
-        # --------------------------------
-
-        self.audit(
-            AuditLog.Events.AGREEMENT_ITEM_DETACHED,
-            target=item,
-            description=(
-                f"{request.user.email} detached asset "
-                f"{asset.public_id} "
-                f"from agreement "
-                f"{agreement.public_id}"
-            ),
-            metadata={
-                "agreement_public_id":
-                    agreement.public_id,
-
-                "agreement_name":
-                    agreement.name,
-
-                "asset_public_id":
-                    asset.public_id,
-
-                "asset_name":
-                    getattr(asset, "name", ""),
-
-                "asset_type":
-                    item.asset_type,
-
-                "agreement_item_public_id":
-                    item.public_id,
-
-                "performed_by":
-                    request.user.email,
-            },
-        )
-
-
-        item.delete()
-
-        return Response(
-            {
-                "detail":
-                "Asset detached from agreement."
-            },
-            status=status.HTTP_200_OK,
-        )
-    
-class AgreementHistoryViewSet( ScopeFilterMixin, viewsets.ReadOnlyModelViewSet, ):
-
-    queryset = (
-        AgreementHistory.objects
-        .select_related(
-            "agreement",
-            "user",
-        )
-        .order_by("-created_at")
-    )
-
-    # permission_classes = [AssetAgreementPermission]
-
-    serializer_class = AgreementHistorySerializer
-
-    pagination_class = FlexiblePagination
-
-    lookup_field = "id"
-
-class AgreementItemHistoryViewSet(
-    ScopeFilterMixin,
-    viewsets.ReadOnlyModelViewSet,
-):
-
-    queryset = (
-        AgreementItemHistory.objects
-        .select_related(
-            "agreement",
-            "agreement_item",
-            "user",
-        )
-        .order_by("-created_at")
-    )
-
-    # permission_classes = [AssetAgreementPermission]
-
-    serializer_class = AgreementItemHistorySerializer
-
-    pagination_class = FlexiblePagination
-
-    lookup_field = "id"
